Classes and Objects
Create a class Box that uses a  parameterized constructor to initialize the dimensions of a box. The dimensions of the box are width, height & depth. The class should have a method that can return the volume of the box. Create an object of the Box class and test the functionalities.